@@ -4,6 +4,8 @@ class ToolbarLayoutSpec {
   static const double height = 56;
   static const double iconSize = 28;
   static const Color backgroundColor = Color(0xFF232323);
-  static const EdgeInsets padding =
-      EdgeInsets.symmetric(horizontal: 12, vertical: 8);
+  static const EdgeInsets padding = EdgeInsets.symmetric(
+    horizontal: 12,
+    vertical: 8,
+  );
 }
