@@ -1,13 +1,21 @@
 import 'package:flutter/material.dart';
 
 class TypographyScale {
-  static const TextStyle headline =
-      TextStyle(fontSize: 32, fontWeight: FontWeight.bold);
-  static const TextStyle title =
-      TextStyle(fontSize: 20, fontWeight: FontWeight.bold);
+  static const TextStyle headline = TextStyle(
+    fontSize: 32,
+    fontWeight: FontWeight.bold,
+  );
+  static const TextStyle title = TextStyle(
+    fontSize: 20,
+    fontWeight: FontWeight.bold,
+  );
   static const TextStyle body = TextStyle(fontSize: 16);
-  static const TextStyle caption =
-      TextStyle(fontSize: 12, color: Colors.white54);
-  static const TextStyle button =
-      TextStyle(fontSize: 16, fontWeight: FontWeight.w600);
+  static const TextStyle caption = TextStyle(
+    fontSize: 12,
+    color: Colors.white54,
+  );
+  static const TextStyle button = TextStyle(
+    fontSize: 16,
+    fontWeight: FontWeight.w600,
+  );
 }
