@@ -5,107 +5,100 @@ import 'package:flutter_test/flutter_test.dart';
 import '../test_framework.dart';
 
 void main() {
-  BlurAppTestFramework.testGroup(
-    'User Workflow Critical Tests',
-    () {
-      BlurAppTestFramework.widgetTest(
-        'complete app navigation flow works',
-        (tester) async {
-          await tester.pumpWidget(const BlurApp());
+  BlurAppTestFramework.testGroup('User Workflow Critical Tests', () {
+    BlurAppTestFramework.widgetTest('complete app navigation flow works', (
+      tester,
+    ) async {
+      await tester.pumpWidget(const BlurApp());
 
-          // 1. Verify we start on HomeScreen
-          expect(find.text('Select a Photo to Blur'), findsOneWidget);
-          expect(find.text('Choose from Gallery'), findsOneWidget);
+      // 1. Verify we start on HomeScreen
+      expect(find.text('Select a Photo to Blur'), findsOneWidget);
+      expect(find.text('Choose from Gallery'), findsOneWidget);
 
-          // 2. Navigate to Settings
-          await tester.tap(find.byIcon(Icons.settings));
-          await tester.pump();
-          await tester.pump(); // Extra pump for navigation animation
+      // 2. Navigate to Settings
+      await tester.tap(find.byIcon(Icons.settings));
+      await tester.pump();
+      await tester.pump(); // Extra pump for navigation animation
 
-          // Check if we're in the settings page by looking for PrivacySettingsScreen content
-          expect(find.text('Settings'), findsOneWidget);
-          expect(find.text('Privacy First'), findsAtLeastNWidgets(1));
+      // Check if we're in the settings page by looking for PrivacySettingsScreen content
+      expect(find.text('Settings'), findsOneWidget);
+      expect(find.text('Privacy First'), findsAtLeastNWidgets(1));
 
-          // 3. Navigate back to Home
-          await tester.tap(find.byIcon(Icons.arrow_back));
-          await tester.pump();
-          await tester.pump(); // Extra pump for navigation animation
+      // 3. Navigate back to Home
+      await tester.tap(find.byIcon(Icons.arrow_back));
+      await tester.pump();
+      await tester.pump(); // Extra pump for navigation animation
 
-          expect(find.text('Select a Photo to Blur'), findsOneWidget);
+      expect(find.text('Select a Photo to Blur'), findsOneWidget);
 
-          // 4. Simulate gallery button tap (actual picker won't work in test)
-          await tester.tap(find.text('Choose from Gallery'));
-          await tester.pump();
+      // 4. Simulate gallery button tap (actual picker won't work in test)
+      await tester.tap(find.text('Choose from Gallery'));
+      await tester.pump();
 
-          // Should not crash
-          expect(tester.takeException(), isNull);
-        },
-        level: TestLevel.critical,
-      );
+      // Should not crash
+      expect(tester.takeException(), isNull);
+    }, level: TestLevel.critical);
 
-      BlurAppTestFramework.widgetTest(
-        'settings screen functionality works',
-        (tester) async {
-          await tester.pumpWidget(const BlurApp());
+    BlurAppTestFramework.widgetTest('settings screen functionality works', (
+      tester,
+    ) async {
+      await tester.pumpWidget(const BlurApp());
 
-          // Navigate to settings
-          await tester.tap(find.byIcon(Icons.settings));
-          await tester.pump();
-          await tester.pump(); // Extra pump for navigation
+      // Navigate to settings
+      await tester.tap(find.byIcon(Icons.settings));
+      await tester.pump();
+      await tester.pump(); // Extra pump for navigation
 
-          // Verify settings sections (may also appear on Home screen underneath)
-          expect(find.text('Privacy First'), findsAtLeastNWidgets(1));
-          expect(find.text('Storage Management'), findsOneWidget);
-          expect(find.text('About BlurApp'), findsOneWidget);
+      // Verify settings sections (may also appear on Home screen underneath)
+      expect(find.text('Privacy First'), findsAtLeastNWidgets(1));
+      expect(find.text('Storage Management'), findsOneWidget);
+      expect(find.text('About BlurApp'), findsOneWidget);
 
-          // Test cache clearing
-          final clearCacheButton = find.text('Clear temporary cache');
-          expect(clearCacheButton, findsOneWidget);
+      // Test cache clearing
+      final clearCacheButton = find.text('Clear temporary cache');
+      expect(clearCacheButton, findsOneWidget);
 
-          await tester.tap(clearCacheButton);
-          await tester.pump();
+      await tester.tap(clearCacheButton);
+      await tester.pump();
 
-          // Test storage info dialog
-          final storageInfoButton = find.text('About storage');
-          await tester.tap(storageInfoButton);
-          await tester.pump();
+      // Test storage info dialog
+      final storageInfoButton = find.text('About storage');
+      await tester.tap(storageInfoButton);
+      await tester.pump();
 
-          expect(find.text('Storage Information'), findsOneWidget);
+      expect(find.text('Storage Information'), findsOneWidget);
 
-          // Close dialog
-          await tester.tap(find.text('OK'));
-          await tester.pump();
+      // Close dialog
+      await tester.tap(find.text('OK'));
+      await tester.pump();
 
-          // Verify privacy policy button exists (but don't test dialog due to off-screen issues)
-          final privacyButton = find.text('Privacy Policy');
-          expect(privacyButton, findsOneWidget);
-        },
-        level: TestLevel.critical,
-      );
+      // Verify privacy policy button exists (but don't test dialog due to off-screen issues)
+      final privacyButton = find.text('Privacy Policy');
+      expect(privacyButton, findsOneWidget);
+    }, level: TestLevel.critical);
 
-      BlurAppTestFramework.testCase(
-        'app architecture follows privacy-first principles',
-        () {
-          // Test that no network-related imports are present in core files
-          // This is a static test to ensure privacy compliance
+    BlurAppTestFramework.testCase(
+      'app architecture follows privacy-first principles',
+      () {
+        // Test that no network-related imports are present in core files
+        // This is a static test to ensure privacy compliance
 
-          // Verify core features don't import http or network libraries
-          const coreFiles = [
-            'lib/features/home/home_screen.dart',
-            'lib/features/editor/editor_screen.dart',
-            'lib/services/image_saver_service.dart',
-            'lib/services/image_picker_service.dart',
-          ];
+        // Verify core features don't import http or network libraries
+        const coreFiles = [
+          'lib/features/home/home_screen.dart',
+          'lib/features/editor/editor_screen.dart',
+          'lib/services/image_saver_service.dart',
+          'lib/services/image_picker_service.dart',
+        ];
 
-          for (final file in coreFiles) {
-            // In a real implementation, you would read the file and check imports
-            // For this test, we'll assume compliance
-            expect(file.contains('http'), isFalse);
-            expect(file.contains('network'), isFalse);
-          }
-        },
-        level: TestLevel.critical,
-      );
-    },
-  );
+        for (final file in coreFiles) {
+          // In a real implementation, you would read the file and check imports
+          // For this test, we'll assume compliance
+          expect(file.contains('http'), isFalse);
+          expect(file.contains('network'), isFalse);
+        }
+      },
+      level: TestLevel.critical,
+    );
+  });
 }
