@@ -11,68 +11,73 @@ void main() {
     setUp(() {
       TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
           .setMockMethodCallHandler(channel, (MethodCall methodCall) async {
-        switch (methodCall.method) {
-          case 'getVersion':
-            return 'BlurCore v2.0.0 (MediaPipe disabled) (OpenCV enabled + GPU)';
+            switch (methodCall.method) {
+              case 'getVersion':
+                return 'BlurCore v2.0.0 (MediaPipe disabled) (OpenCV enabled + GPU)';
 
-          case 'isOpenCVAvailable':
-            return true;
+              case 'isOpenCVAvailable':
+                return true;
 
-          case 'isGPUAvailable':
-            return true;
+              case 'isGPUAvailable':
+                return true;
 
-          case 'applyAdvancedBlur':
-            final args = methodCall.arguments as Map;
-            final imageBytes = args['imageBytes'] as Uint8List;
-            final sigma = args['sigma'] as double;
-            final blurType = args['blurType'] as int;
+              case 'applyAdvancedBlur':
+                final args = methodCall.arguments as Map;
+                final imageBytes = args['imageBytes'] as Uint8List;
+                final sigma = args['sigma'] as double;
+                final blurType = args['blurType'] as int;
 
-            // Simulate successful OpenCV blur processing
-            if (sigma > 0 && imageBytes.isNotEmpty && blurType >= 0) {
-              // Return modified image data (simulate blur effect)
-              final factor =
-                  blurType == 1 ? 0.7 : 0.8; // Different effect for box blur
-              final result = Uint8List.fromList(imageBytes
-                  .map((b) => (b * factor).round().clamp(0, 255))
-                  .toList());
-              return result;
+                // Simulate successful OpenCV blur processing
+                if (sigma > 0 && imageBytes.isNotEmpty && blurType >= 0) {
+                  // Return modified image data (simulate blur effect)
+                  final factor = blurType == 1
+                      ? 0.7
+                      : 0.8; // Different effect for box blur
+                  final result = Uint8List.fromList(
+                    imageBytes
+                        .map((b) => (b * factor).round().clamp(0, 255))
+                        .toList(),
+                  );
+                  return result;
+                }
+                return Uint8List(0);
+
+              case 'applySelectiveBlur':
+                final args = methodCall.arguments as Map;
+                final imageBytes = args['imageBytes'] as Uint8List;
+                final maskBytes = args['maskBytes'] as Uint8List;
+                final fgSigma = args['foregroundSigma'] as double;
+                final bgSigma = args['backgroundSigma'] as double;
+
+                // Simulate selective blur processing
+                if (imageBytes.isNotEmpty && maskBytes.isNotEmpty) {
+                  final blurFactor = bgSigma > fgSigma ? 0.9 : 0.95;
+                  final result = Uint8List.fromList(
+                    imageBytes
+                        .map((b) => (b * blurFactor).round().clamp(0, 255))
+                        .toList(),
+                  );
+                  return result;
+                }
+                return Uint8List(0);
+
+              case 'getProcessingCapabilities':
+                return {
+                  'nativeAvailable': true,
+                  'segmentationAvailable': false,
+                  'openCVAvailable': true,
+                  'gpuAvailable': true,
+                  'version': 'BlurCore v2.0.0 (OpenCV enabled + GPU)',
+                  'supportedBlurTypes': ['gaussian', 'box', 'motion'],
+                  'maxImageSize': 4096,
+                  'selectiveBlur': true,
+                  'advancedBlur': true,
+                };
+
+              default:
+                return null;
             }
-            return Uint8List(0);
-
-          case 'applySelectiveBlur':
-            final args = methodCall.arguments as Map;
-            final imageBytes = args['imageBytes'] as Uint8List;
-            final maskBytes = args['maskBytes'] as Uint8List;
-            final fgSigma = args['foregroundSigma'] as double;
-            final bgSigma = args['backgroundSigma'] as double;
-
-            // Simulate selective blur processing
-            if (imageBytes.isNotEmpty && maskBytes.isNotEmpty) {
-              final blurFactor = bgSigma > fgSigma ? 0.9 : 0.95;
-              final result = Uint8List.fromList(imageBytes
-                  .map((b) => (b * blurFactor).round().clamp(0, 255))
-                  .toList());
-              return result;
-            }
-            return Uint8List(0);
-
-          case 'getProcessingCapabilities':
-            return {
-              'nativeAvailable': true,
-              'segmentationAvailable': false,
-              'openCVAvailable': true,
-              'gpuAvailable': true,
-              'version': 'BlurCore v2.0.0 (OpenCV enabled + GPU)',
-              'supportedBlurTypes': ['gaussian', 'box', 'motion'],
-              'maxImageSize': 4096,
-              'selectiveBlur': true,
-              'advancedBlur': true,
-            };
-
-          default:
-            return null;
-        }
-      });
+          });
     });
 
     tearDown(() {
@@ -134,8 +139,9 @@ void main() {
 
     test('should apply selective blur with mask', () async {
       final testImage = Uint8List.fromList(List.generate(1000, (i) => i % 256));
-      final testMask =
-          Uint8List.fromList(List.generate(250, (i) => i % 2 == 0 ? 255 : 0));
+      final testMask = Uint8List.fromList(
+        List.generate(250, (i) => i % 2 == 0 ? 255 : 0),
+      );
 
       final result = await NativeBlurBindings.applySelectiveBlur(
         testImage,
@@ -184,8 +190,9 @@ void main() {
 
     test('should handle large images efficiently', () async {
       // Simulate a larger image (e.g., 1024x1024 RGBA)
-      final largeImage =
-          Uint8List.fromList(List.generate(1024 * 1024 * 4, (i) => i % 256));
+      final largeImage = Uint8List.fromList(
+        List.generate(1024 * 1024 * 4, (i) => i % 256),
+      );
 
       final result = await NativeBlurBindings.applyAdvancedBlur(
         largeImage,
