@@ -17,100 +17,97 @@ import '../test_framework.dart';
 /// - Never be skipped in CI
 
 void main() {
-  BlurAppTestFramework.testGroup(
-    'Core Feature Tests',
-    () {
-      // Example: Core blur algorithm test
-      BlurAppTestFramework.testCase(
-        'blur algorithm produces valid output for all supported strengths',
-        () {
-          // Test core business logic
-          for (int strength = 1; strength <= 50; strength++) {
-            BlurAppAssertions.assertValidBlurParams(strength);
+  BlurAppTestFramework.testGroup('Core Feature Tests', () {
+    // Example: Core blur algorithm test
+    BlurAppTestFramework.testCase(
+      'blur algorithm produces valid output for all supported strengths',
+      () {
+        // Test core business logic
+        for (int strength = 1; strength <= 50; strength++) {
+          BlurAppAssertions.assertValidBlurParams(strength);
 
-            // Test blur calculation logic
-            final result = _mockBlurCalculation(strength);
-            expect(result, isNotNull);
-            expect(result, greaterThan(0));
-          }
-        },
-        level: TestLevel.core,
-      );
+          // Test blur calculation logic
+          final result = _mockBlurCalculation(strength);
+          expect(result, isNotNull);
+          expect(result, greaterThan(0));
+        }
+      },
+      level: TestLevel.core,
+    );
 
-      // Example: Data validation test
-      BlurAppTestFramework.testCase(
-        'image data validation rejects invalid inputs',
-        () {
-          // Test null input
-          expect(() => _validateImageData(null), throwsArgumentError);
+    // Example: Data validation test
+    BlurAppTestFramework.testCase(
+      'image data validation rejects invalid inputs',
+      () {
+        // Test null input
+        expect(() => _validateImageData(null), throwsArgumentError);
 
-          // Test empty input
-          expect(() => _validateImageData([]), throwsArgumentError);
+        // Test empty input
+        expect(() => _validateImageData([]), throwsArgumentError);
 
-          // Test invalid format
-          expect(() => _validateImageData([1, 2, 3]), throwsArgumentError);
-        },
-        level: TestLevel.core,
-      );
+        // Test invalid format
+        expect(() => _validateImageData([1, 2, 3]), throwsArgumentError);
+      },
+      level: TestLevel.core,
+    );
 
-      // Example: Essential calculation test
-      BlurAppTestFramework.testCase(
-        'coordinate transformation maintains precision',
-        () {
-          final testCases = [
-            {
-              'input': [0, 0],
-              'scale': 2.0,
-              'expected': [0, 0]
-            },
-            {
-              'input': [100, 100],
-              'scale': 0.5,
-              'expected': [50, 50]
-            },
-            {
-              'input': [255, 255],
-              'scale': 1.0,
-              'expected': [255, 255]
-            },
-          ];
+    // Example: Essential calculation test
+    BlurAppTestFramework.testCase(
+      'coordinate transformation maintains precision',
+      () {
+        final testCases = [
+          {
+            'input': [0, 0],
+            'scale': 2.0,
+            'expected': [0, 0],
+          },
+          {
+            'input': [100, 100],
+            'scale': 0.5,
+            'expected': [50, 50],
+          },
+          {
+            'input': [255, 255],
+            'scale': 1.0,
+            'expected': [255, 255],
+          },
+        ];
 
-          for (final testCase in testCases) {
-            final input = testCase['input'] as List<int>;
-            final scale = testCase['scale'] as double;
-            final expected = testCase['expected'] as List<int>;
+        for (final testCase in testCases) {
+          final input = testCase['input'] as List<int>;
+          final scale = testCase['scale'] as double;
+          final expected = testCase['expected'] as List<int>;
 
-            final result = _transformCoordinates(input, scale);
-            expect(result, equals(expected));
-          }
-        },
-        level: TestLevel.core,
-      );
+          final result = _transformCoordinates(input, scale);
+          expect(result, equals(expected));
+        }
+      },
+      level: TestLevel.core,
+    );
 
-      // Example: State consistency test
-      BlurAppTestFramework.testCase(
-        'core state transitions maintain data integrity',
-        () {
-          final state = _createTestState();
+    // Example: State consistency test
+    BlurAppTestFramework.testCase(
+      'core state transitions maintain data integrity',
+      () {
+        final state = _createTestState();
 
-          // Test initial state
-          expect(state.isValid, isTrue);
-          expect(state.hasChanges, isFalse);
+        // Test initial state
+        expect(state.isValid, isTrue);
+        expect(state.hasChanges, isFalse);
 
-          // Test state modification
-          state.applyChange('test change');
-          expect(state.hasChanges, isTrue);
-          expect(state.isValid, isTrue);
+        // Test state modification
+        state.applyChange('test change');
+        expect(state.hasChanges, isTrue);
+        expect(state.isValid, isTrue);
 
-          // Test state reset
-          state.reset();
-          expect(state.hasChanges, isFalse);
-          expect(state.isValid, isTrue);
-        },
-        level: TestLevel.core,
-      );
-    },
-  );
+        // Test state reset
+        state.reset();
+        expect(state.hasChanges, isFalse);
+        expect(state.isValid, isTrue);
+      },
+      level: TestLevel.core,
+    );
+  });
 }
 
 // Mock implementations for example tests
