@@ -4,14 +4,11 @@ import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group('HomeScreenMVP Widget Tests', () {
-    testWidgets('should display app title and buttons',
-        (WidgetTester tester) async {
+    testWidgets('should display app title and buttons', (
+      WidgetTester tester,
+    ) async {
       // Build our app and trigger a frame
-      await tester.pumpWidget(
-        const MaterialApp(
-          home: HomeScreenMVP(),
-        ),
-      );
+      await tester.pumpWidget(const MaterialApp(home: HomeScreenMVP()));
 
       // Verify app title is displayed
       expect(find.text('BlurApp'), findsOneWidget);
@@ -23,30 +20,25 @@ void main() {
 
       // Verify privacy section is displayed
       expect(find.text('Privacy First'), findsOneWidget);
-      expect(find.textContaining('All processing happens on your device'),
-          findsOneWidget);
+      expect(
+        find.textContaining('All processing happens on your device'),
+        findsOneWidget,
+      );
     });
 
     testWidgets('should have app icon', (WidgetTester tester) async {
       // Build our app and trigger a frame
-      await tester.pumpWidget(
-        const MaterialApp(
-          home: HomeScreenMVP(),
-        ),
-      );
+      await tester.pumpWidget(const MaterialApp(home: HomeScreenMVP()));
 
       // Verify app icon is displayed
       expect(find.byIcon(Icons.blur_on), findsOneWidget);
     });
 
-    testWidgets('should have gallery and camera buttons',
-        (WidgetTester tester) async {
+    testWidgets('should have gallery and camera buttons', (
+      WidgetTester tester,
+    ) async {
       // Build our app and trigger a frame
-      await tester.pumpWidget(
-        const MaterialApp(
-          home: HomeScreenMVP(),
-        ),
-      );
+      await tester.pumpWidget(const MaterialApp(home: HomeScreenMVP()));
 
       // Verify gallery button with icon
       expect(find.byIcon(Icons.photo_library), findsOneWidget);
@@ -61,11 +53,7 @@ void main() {
 
     testWidgets('buttons should be tappable', (WidgetTester tester) async {
       // Build our app and trigger a frame
-      await tester.pumpWidget(
-        const MaterialApp(
-          home: HomeScreenMVP(),
-        ),
-      );
+      await tester.pumpWidget(const MaterialApp(home: HomeScreenMVP()));
 
       // Find the gallery button and verify it can be tapped
       final galleryButton = find.text('Choose from Gallery');
@@ -79,14 +67,11 @@ void main() {
       // trigger image picker which requires platform mocking
     });
 
-    testWidgets('should have proper layout structure',
-        (WidgetTester tester) async {
+    testWidgets('should have proper layout structure', (
+      WidgetTester tester,
+    ) async {
       // Build our app and trigger a frame
-      await tester.pumpWidget(
-        const MaterialApp(
-          home: HomeScreenMVP(),
-        ),
-      );
+      await tester.pumpWidget(const MaterialApp(home: HomeScreenMVP()));
 
       // Verify main structure elements
       expect(find.byType(Scaffold), findsOneWidget);
